Change visuals and fix spacing of text and divs.
I added in placeholder images for the rest of the cards so that I had an idea of how things might look put together. I also smoothed out the color scheme chosen and refined the size of my div containers. Additionally I added more text as a placeholder and gave it the desired text decorations.